Move files to components
