+++ eq-author/src/App/page/PropertiesPanel/GroupedAnswerProperties/Fallback.js
@@ -2,8 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
 import CustomPropTypes from "custom-prop-types";
-import InlineField from "./InlineField";
-import MultiLineField from "./MultiLineField";
+import InlineField from "components/AdditionalContent/AnswerProperties/Format/InlineField";
+import MultiLineField from "components/AdditionalContent/AnswerProperties/Format/MultiLineField";
 import ToggleProperty from "components/AdditionalContent/ToggleProperty";
 import { SimpleSelect } from "components/Forms/Select";
 import Collapsible from "components/Collapsible";
