+++ eq-author/src/components/AdditionalContent/AnswerProperties/Answers/DurationProperties.test.js
