+++ eq-author/src/components/AdditionalContent/AnswerProperties/Answers/TextProperties.js
