+++ eq-author/src/components/AdditionalContent/AnswerProperties/Answers/index.js
@@ -1,4 +1,4 @@
-export { default as Decimal } from "../../Decimal";
+export { default as Decimal } from "../Decimal";
 export { default as DateFormat } from "./DateFormat";
 export { default as UnitProperties } from "./UnitProperties";
 export { default as DurationProperties } from "./DurationProperties";
