+++ eq-author/src/components/AdditionalContent/AnswerProperties/Answers/updateAnswersOfTypeMutation.graphql
