+++ eq-author/src/components/AdditionalContent/AnswerProperties/Decimal/index.js
